requirements.txt
streamlit>=1.33.0
llama-index
llama-index-llms-google-genai
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
chromadb>=0.4.0
docling>=2.80.0
rank_bm25==0.2.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests>=2.32.5
google-genai
pydantic>=2.0.0
supabase
